requirements.txt
pandas~=2.2.3
numpy~=2.1.2
matplotlib~=3.9.2
opencv-python~=4.10.0.84
scipy~=1.14.1
pillow~=11.0.0